Details in Description
I removed the static size values of the canvas, sidebar, and stuff from HTML/CSS, and replaced it with CSS grid, this way it's much more adjustable and easier to maintain/oversee.

Also, a sidebar has been added, we can later add the user's skills here, categorized, and details can be added on hover event.