today in desc
fixed tree searching, now can add trees
added a couple of sample trees to play with
added user searching, can get the public  user data of others, dont do anything with this yet, this feature needs the new design.
The user's trees are now dynamically listed on the sidebar, and can be switched.